simple fix to catastrophic crash
If the user gets the file address wrong, or a file gets deleted without realising, it causes a full crash of any kernel (e.g. jupyter notebook). This has caused some headaches in the past. This is a simple fix that allows the user to know something has gone wrong but without causing the kernel to crash.